added homework three files
we had to implement a queue using a double linked list class and make a function to check if parenthesis were balanced in an expression(parenthesis balancing) using a stack. also we had to implement codes of two examples  of real life example of queue and a stack. for a stack i made a task manager class and for priority queue i made a ambulance dispatch system